DocAI/backups/DocAI_WORKING_2025-06-21_02-02-15_FULL_INTEGRATION/requirements.txt
# Flask and Web Framework
flask>=2.3.3
flask-cors>=4.0.1
gunicorn>=20.1.0
Werkzeug>=2.3.7

# Environment and Configuration
python-dotenv>=1.0.0

# AI and Language Models
groq>=0.10.0
openai>=1.3.7
google-generativeai>=0.3.0
anthropic>=0.7.0
pocketgroq>=0.5.6

# PyTorch and ML
torch>=2.2.0
transformers>=4.35.0

# NLP and Text Processing
nltk>=3.8.1
sentence-transformers>=2.2.2

# LangChain Ecosystem
langchain-core>=0.0.10
langchain-community>=0.0.10
langchain-groq>=0.0.1
langchain-huggingface>=0.1.1

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.0.1
docx2txt>=0.8
PyMuPDF>=1.23.0
pdf2docx>=0.5.8

# Vector Database and RAG
faiss-cpu>=1.7.4

# Web Scraping and Processing
beautifulsoup4>=4.12.2
requests>=2.31.0

# Data Validation and Serialization
pydantic>=1.10.13

# Image Processing
Pillow>=10.0.0

# Development and Debugging (Optional)
streamlit>=1.28.0

# Security and Environment
cryptography>=41.0.0

# Testing (Optional - for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Additional utilities that might be needed
numpy>=1.24.0
pandas>=2.0.0

# File and Data Handling
pathlib2>=2.3.7; python_version<"3.6"
